GLTextFontTexture: Finish copyGlyphImageToTexture. Start testing
Implement exportTextureBitmaps() for all classes up to GLTextFontCache and
call it from the main program after all rendering.
It writes a .data file per texture with font name, size, and number pixels in
x and y in the file name. You can open it with Gimp, providing the format
 (gray 8-bit and width and height according to the file name.

Note that the image is upside-down because GL textures build up bottom to top
whereas Gimp builds an image like the screen is written top to bottom.

ToDo: When last characters of the current line do not have a character below
from the previous line (i.e. the previous line ends left of the current
character) this character will drop to the bottom, thus overwriting a
character at a lower line at the bottom.