GLTextFontCacheItem: Add addGlyphAsTofu() and use it
When a glyph cannot be loaded or cannot be rendered add it to the font
cache with the coordinates (and the image) of the tofu glyph (index 0)
todo: The absolute emergency, that even the tofu glyph with index 0
cannot be loaded and rendered is not yet covered.

GLTextGlyphBBox: Declare inline functions as const where possible.