Build system: Build with sytem pango.
The sub-project 3rdParty/pango is version 1.55.1 which requires a newer
harfbuzz version than Ubuntu Noble ships with.
I am just too lazy to include also Harfbuzz as sub-project.
I included pango in the project mostly to be able to debug it for better
understanding a large and complex library.
Now I am happy with the OS supplied version.