GLTextGlyphBBox: Refactor and make it a template class
Change GLTextGlyphBBox to a template class GLTextGlyphBBoxT.
Refactor the original GLTextGlyphBBox to GLTextGlyphBBoxI as
implementation of the template with type T = int32_t.
Add implementation alias GLTextGlyphBBoxF for floats.

Add a normalized BBox texturePositionNormalized to
GLTextFontCacheGlyphItem, and calculate the normalized texture position
from the absolute positon, and the texture size.