main: Change apterture angle and near and far planes to make the coorinate units 1:1 per pixel
Apterture angle is set 30deg, and the near plane is identical to the window
width. The 0-z plane is 2*windowWidth, and the fare plane is 3*windowWidth.
Thus the frustum is the window width in all directions.
One unit in client space is equal to one pixel in real life at the plane of 0 z. For z-values unequal 0 of course the perspecitve shrink or stretch apply.

Adjust the dimensions of the test renderers accordinly. The texture renderer is
now 512x512 units wide which is the original dimension of the png image.