configure.ac, */Makefile.am: Change SDL build to shared library
Using SDL3 as static library does lead to inexplicable crashes in many commits
Even when the program does not crash right away X11 and wayland video drivers
produce a lot of errors, not able to resolve lots of szstem library symbols;
wayland does not open a visible window.

Configure and make require two passes:
The first is to configure and build SDL3 in 3rdParty/SDL/, and install it
locally, the second pass is to configure my sources in src/ to use the just
build SDL via pkg-config.
PKG_CONFIG_PATH is set in configure before checking modules to find the
locally installed SDL3.