configure.ac: Improve building and installing submodules
Configure will configure, build, and install the submodules Pango and SDL3
when no suitable packages can be detected.
The further configuration of the include paths and library and linker options
for the main program depend on exiting submodule packages.
The subpackages must be linked as shared libs.
The attempt to staticically link particularly SDL3 failed miserably.
It linked fine but depending on the git commit of SDL it crashed randomly in
different parts of the program, even when up the point of crash everything
was Valgrind-proof.