GLTextRenderer works!
Primary error was in the test program: I still need a sensible view matrix;
the identity matrix is not a valid view matrix!

ToDo: Use interpolation for the glyhph texture, not nearest.
Apply alpha blending, and use the texture for the alpha channel.
Draw a bounding rectangle below the text box for non-transparent text.