main(): Tests, cleanup, dynamic calculation of aperture angle.
Assume that the camera looks to the origin from the z-axis. Then the aperture
angle is calculated that one coordinate unit is exactly one pixel at the
x-y plane at z-0.
Thus texts are displayed as sharp as rendered into the glyph textures.

Various tests are conducted:
- Non-square aspect ratio (640x480): Display is OK, aperture angle is correct.
- Texts with multiple fonts in one string, and multiple textures per font
 tested: OK
- Resolution of texts at z=0 tested: Identical to the glyph textures.

The render loop now runs endlessly but you can close the window.

Deleted other text rendering which were used for testing the GLTextRenderer
background functionality but were never used in the render loop.