Revert b3e95cef0b583f51757427026171da0bc2aa6115, and make linking working.
Either gcc as linker driver or ld itself seem to ignore multiple listings of
explicit library names to resolve circular dependencies in static libraries.
To "cheat" the linker stage I am listing repeated library ocurrences with "./"
before the actual library path and name.
gcc and ld are not intelligent enough to equate xxx/yyy.a with ./xxx/yyy.a.
Thus this hack works.