CirclePolygonVertexContainer: Correct order of vertexes, z-value of the normal
Initialize the z-value of the normals to 1.0. I can adjust the actual z-value
by a factor in the internal model matrix. This would not be possible with a
value 1.0.

Start the vertex list always with the secondary vertex. The circle is always
drawn counter-clock wise. This way the triangles are also drawn counter-clock
wise too, and the normal vertex matches the GL normal.