GLTextRenderer::draw: Draw overlapping Glyphs completely
Partially overlapping glyphs were cut off where an overlap occurred.
Occurrences are:
- Overlap due to kerning in western scripts like combinations of "VA",
"//" "Va"
- Complex scripts like Thai where diacritics are partially overlapping
- Calligraphic scripts like Arabic Naskh or Bengali with connected
glyphs showed gaps in the flow of the script.
Set the depth test function to "less or equal" instead of the default
function "less" which leads to the suppression of overlapping parts of
the subsequent glyph.
Partially overwriting of glzphs is no issue, contrary it is desired. All
glyphs are rendered in transparent mode, i.e. the background of a glyph
is completely transparent. The rendering of glyphs was intentionally
implemented this way to allow for undercuts and overlaps of glyphs.
For opaque text a background rectangle is rendered before the rendering
of the glyphs in transparent mode.
After rendering text the original depth function is restored.