CirclePartialArcRenderer: Complete implementation for coarse arc.
Implementation for the coarse arc is complete.
This means the implementation draws an arc of the given circle segments
which stay within the arc range.
However the exact arc range is not reached due to the coarse nature of
the polygon segments which make up a circle or arc.
For my tests a full circle is comprised of 64 segments, i.e. the arc has
a stepping of 5.625 degrees.

Next to come is adding a segment to fill the arc exactlty to the arc's
angle.