CirclePartialArcRenderer: Draw additional segement to complete the arc.
Draw one additional segment of the polygon, but it start at such an angle that
it ends exactly at the designated arc end.
Using the fixed segments of the polygon has two draw-backs:
1. For arc ranges smaller than the polygon increment the single additional
segment will under-shoot the start of the arc since the calculation lets the
singdle end segment end exactly at the end of that arc.
2. In other cases than 1. the additional arc partially overlaps the coarse arc.
this overlap is clearly visible for transparent arcs as the alpha value of the
object color add up by the overlapping drawing.
So, the final solution will be to make up a custom segment which exactly fills
the gap between the end of the coarse arc from the fixed polygon and the
designed arc end.

normalizeAngle(): Implement the method only once but as a template function
which will instanciate itself for any angle type (degree or radiants).
Comparisons are against 0 or full circle which both are available for the
template too.