CirclePartialArcRenderer: Complete and test the implmentation.
Add an array of 4 vertexes. Fill the array with the vertex pair
of the last segment of the coarse polygon. Calculate x and y of
the last 2 vertexes from the arc end angle.
Draw the 2 triangles as a strip as the end segment which fills
the gap between the end of the coarse polygon and the exact end of the
arc angle.