resources/: Extend the build system to incorporate binary files into the program
Create rules to add .png or (future) .jpg and jpeg in the list of sources
for a static resources library.
These files are included per assembler into an object .o file.
In addition per binary file a header file is generated to permit access
to the memory location and length of the included resource.