SquareTextureRenderer(): Revert to the original coordinates.
Revert the initialization of the coordinates in the constructor to the values
as in commit 8cc648c0dfde75e8a6756d6bcf43c403d15ddd4f.
Obviously the orcer of vertexes was not correct to form a triangle strip in the
intended form of two adjacent triangles forming a rectangle.