Change compiler standard to C++ 20.
Now I can use std::source_location to pass information about the function
and the source location to an error handler in a comprehensive way.

One side effect is that fmt::format_string expects to be constructed
from either a string literal or a constexpr string.
I need to wrap the translated, dynamic format strings with fmt::runtime().