Uuid.h: Add constexpr functions strToUuidBinary and checkUuidString.
strToUuidBinary can be used to create constexpr Uuid::UuidBinaryT objects
directly from the UUID string which you can produce for example with
the program uuidgen.

checkUuidString can be used to check at compile time if constexpr UUID string
is conforming to the format like:
9d90ebdb-239a-49cf-877e-be334e4d6590
with the '-' characters in the right place, and the other characters hex digits
and the entire string being 36 characters long.