Add new class OevControls::RootControl.
I solved the trickiest part: Pass a weak pointer to the constructor
which points to itself.
I solved it by using std::allocator and separating allocation of memory
and calling the constructor separately, in the meantime constructing
the owning shared_ptr to the (still uninitialized, but existing) memory,
and passing that as a weak_ptr to the constructor.