RenderBase - RenderStandardUniforms: Re-design and finalize the API.
- Remove external access to the shared pointer instances.
- All shared pointers are valid at any time
- A new objct is created with all shared pointers with valid and generic
initial data.
- Shared pointers for Matrix data are in a struct of the matrix and an update
sequence counter.
- Any writable access to Matrixes increments the global change sequence for
the matrix in the shared pointer.
- Calculation of MVMatrix and MVP matrix is lazy, and driven by the sequence
number of underlying matrixes. When read access to MVMatrix or MVP matrix is
requested it is automatically re-calculated when the change sequence of
the parent matrixes is higher than the own change sequence.

ToDo: For access to MVPMatrix also the model and view matrix change sequences
must be checked.