GLFramework: Add event handling method. Retrieve window ID where possible.
This is the start of the event handling distribution with the even handler
method. Where possible, and where it makes sense for me at the moment
the method extracts the windowID, and tries to find the renderSurface and
Window.

ToDo: Next is forwarding window related events to the respective renderSurface
and moving the entire main loop from the test implementation in the main module
to GLFramework.