src/Makefile.am: Use a proper library group to link static libraries with circular references.
Automake does not accept linker options in OpenVarioFront_LDADD but only
library names (incl. -lLib).

A library group implies ld options which are not directly accepted.
Therefore I need to cheat automake:
The Makefile.am must not know the list, not even indirectly as variables which
are defined within the Makefile.am.
The definition of the library group with the options --start-group and
--end-group are placed in configure.ac, and funneled into the makefiles with
AC_SUBST().
When I run autoreconf, and with that automake, it does not know the content
of the variable $(LOCAL_LIBS_GROUP) in Makefile.am.
The values are in configure.ac.
They are populated in Makefile when I run configure.
This does the trick.