configure.ac, src/Makefile.am: Make LOCAL_LIBS_GROUP static and local to Makefile.am
Defining LOCAL_LIBS_GROUP in configure.ac as a linker options containing
--start-group - --end-group did work great to resolving circular references
within the libraries.
However since the libraries are not known as list of static lib names in
Makefile.am there was no dependency established between the final execuatble
OpenVarioFront and the static libraries.

Revert to a static list, but list all libs twice.
Once as xx/yy.a and once as ./xx/yy.a. This suffices to trick the toolchain
not ignoring the second occurrence of the lib name when linking, thus
resolving the simple circular references.