ControlBase: Add getRenderUniforms(); use the local matrix also for stretching to size.
Initialize renderUniforms from the parent's one, and immediately reset the
model matrix, and set it to Identity.

Set the scaling factors of locControlModelMatrix when setting the size with
either setPosition() or setTopRight().

Move all code out of onPositionChanged() and onSizeChanged(), in order not to
make an overrider do stuff again that I am doing here (and will probably
forget it anyway).