GLObjectWrappers: Remove bool operators. They are implicitly converted to GLuint!
Replace the bool operators with a valid() method.
The implicit converion of bool to GLuint would allow using the object
e.g. in glBind() passing a bool as an invalid GL handle.

Add reset() to bind objects to allow re-binding a buffer or vertex array
while preseving the original backup buffer or vertex handle.