ControlBase: Correct initialization of parentModelMatrixPtr and renderUniforms.
Initialization was pretty messed up in the constructor.
Now renderUniforms is initialized in the initializer list
by a lambda. The lambda returns either a copy of the parent's renderUniforms
when the control is not a root control, or a newly constructed object
when this is a root control.

parentModelMatrixPtr is initialized empty.
It is set in the body of the constructor to a unity matrix for when this is a root control
It is copied from the renderUniforms when this is not a root control. Here
renderUniforms.modelMatrix is reset to a unity matrix (now owned by this) when
this is not a RootControl.